还没开始今天的工作 先简单修改下cateen的地图路径为云存储库地图 使得本地代码包大小急剧缩减
