--- cloudfunctions/search_test/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "search_test",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "wx-server-sdk": "~2.3.2"
-  }
-}
