+++ miniprogram/pages/canteen/canteen.wxml
@@ -1,9 +1,9 @@
 <!--pages/Restaurant1/Restaurant1.wxml-->
 <view class="map">
-  <image class="mapitem" src="../../images/canteen/map1.png" ></image>
+  <image class="mapitem" src="https://7375-sutest2020-ukd9i-1304294060.tcb.qcloud.la/images/canteen/map1.png?sign=06bab4fc69da643186505fbd25865289&t=1606363439" ></image>
 </view>
 <view class="store">
     <view class="store_each" wx:for="{{array}}" wx:for-item="item">
-        <text bindtap="btnclick" style='writing-mode:vertical-lr;position:fixed; width: {{item.width*2.34}} rpx;height: {{item.height*2.34}}rpx;left:{{item.locationx*2.34}}rpx;top:{{item.locationy*2.34}}rpx;font-size:10px;'>{{item.name}}</text>
+        <text bindtap="btnclick" data-text="{{item.name}}" style='writing-mode:vertical-lr;position:fixed; width: {{item.width*2.34}} rpx;height: {{item.height*2.34}}rpx;left:{{item.locationx*2.34}}rpx;top:{{item.locationy*2.34}}rpx;font-size:10px;'>{{item.name}}</text>
     </view>
 </view>
