+++ miniprogram/pages/homepage/homepage.js
@@ -49,7 +49,7 @@ Page({
             success: function (res) {
                {
                 console.log('检测数据库函数完成调用')
-                console.log(res)
+                console.log(res.result)
                }
              },
              fail: console.error
