+++ miniprogram/pages/shop/shop.wxml
@@ -2,7 +2,7 @@
   <text class="text1">招牌菜</text>
   <swiper indicator-dots="{{indicatorDots}}" 
     autoplay="{{autoplay}}" interval="{{interval}}" duration="{{duration}}">
-    <block wx:for="{{imgUrls}}">
+    <block wx:for="{{imgUrls}}" wx:key="item" >
       <swiper-item>
         <image src="{{item}}" class="slide-image" />
       </swiper-item>
